PmConversationsCard: Show the real data we have, not a spinner.
This loading indicator was added in 06c4de8e1.  At that time, we
didn't persist the `users` data between runs of the app, and so
when we'd just started up we would lack that data and be unable to
properly show conversations, even though we did persist messages.

We fixed that in 70679b7e3 back in 2018, persisting the users along
with other essential data.  Now this screen can work just fine from
cached data; so let it do so.

[greg: extended commit message]

Fixes: #3241