webview build [nfc]: Pull out rsync command into a function.
We're about to add another step here, for copying KaTeX-related
files out of the KaTeX distribution.  Factor out the routine bits
into a shell function, so we can invoke the same function for that
step too.

This also makes a couple of small changes to the rsync command.
These don't affect the existing behavior, but they help abstract the
shared fiddly-rsync bits, which go inside the function, away from
the specific sets of files to copy, which go at the call site.

 * Use -R aka --relative.  This allows a root source directory to be
   specified once, and all other paths in the command to be relative
   to that root.

 * Use rsync filter rules.  This allows "which files from the source
   should be copied" to be specified all in one place, in a
   well-defined format, and without taking arbitrary rsync arguments
   which could interact in surprising ways with those provided by
   the function.