welcome: Revise suggestion for how to start if no account.
If someone hasn't yet created a Zulip account, then most likely
there's a particular Zulip org they want to be using.  Going to
zulipchat.com won't much help them to find that org.  Instead,
tell them to ask someone who might know.

Also update the string for translation... and make this a normal
`Label` rather than `RawLabel`, so the translation can have a shot
at getting used, oops.  Not sure why it was `RawLabel` to begin with,
but evidently I missed that when adding the string for translation
in the first place in bcbaeb841.