docs/build-run: Highlight Node 10.x requirement more.
A good fraction of our build-trouble reports turn out to be an
incompatible version of Node; it appears people aren't seeing this
instruction, or are ignoring it.

Put it in bold, and add some words to make clear we haven't just
neglected to update the instruction.

Also delete "LTS": while 10.x is still LTS (until 2021-04-30, in a
little over a year), it's no longer the *latest* LTS because 12.x is
also now LTS.

Currently Node 12 (except some older versions of it) doesn't work
because of an issue in Metro; that was fixed in Metro v0.57.0, which
will come with RN v0.62.  Details:
  https://chat.zulip.org/#narrow/stream/48-mobile/topic/getting.20errors.20in.20environment.20building/near/807957
So after we take that RN upgrade, perhaps we can move to Node 12.