fetchActions: Fetch server version on /register and include in REALM_INIT.
Await completion of /server_settings along with /register (already
being awaited) in a Promise.all before setting session.loading to
false with initialFetchComplete.

We considered making a new async action creator for each of these
two requests and giving them both responsibility for setting
session.loading to false with initialFetchComplete. But this is
undesirable because we need to wait for both of them to complete,
not just one.

We had a choice of awaiting each of these two requests sequentially,
but this would result in a longer loading period than awaiting a
Promise.all, which will only take as long as the longer of the two
requests. The longer one is almost certainly /register in most
cases, so this change shouldn't actually increase the loading time
at all.