accountsReducer test: Use exampleData; enable type-checking.
Change makeAccount to use a params object, since more params
will be added in a commit later in this series.

Declare variables for the Account objects to avoid the pattern
of setting, e.g., prevState[0].email to the literal
'one@example.com' and expectedState[0].email to the literal
'one@example.com', which forces the test writer to visually confirm
the strings are identical.