MessageList: Remove needless borderWidth: 0 on WebView style.
borderWidth defaults to 0. No visible changes were seen on iOS or
Android after removing it. 32ce2086c, which introduced this, was a
large change that introduced the whole WebView-based message list,
and it doesn't have detail about the need for this small, odd use of
this property, so there's little reason to think it was carefully
considered in the first place.