connect [nfc]: Extract SelectorProps from Props in 10 files, for clarity.
Also, remove some GlobalState annotations on the state param of the
lambda passed to connectFlowFixMe. Removing them doesn't break
anything, but it removes some type clarity that will be restored in
the upcoming series of commits when `connect` is used instead of
`connectFlowFixMe`.

This, and the upcoming series of commits, fixing
`connectFlowFixMe`s, isn't exhaustive; we're addressing places where
easy fixes were discovered during work on #3765.