narrow [nfc]: Factor out pmNarrowByString.
We have two "constructors" `privateNarrow` and `groupNarrow` for
1:1 PMs and group PMs respectively, but they end up building
basically the same data structure.  Then in a lot of places in our
code we have control flow to handle the group-PM case on the one
hand and the 1:1 case on the other, and it ends up just saying the
same thing twice to treat them both the same way.  (In a few places,
like `getNarrowFromMessage`, we have no such logic and already
implicitly rely on the fact that they're the same underneath.)

In general the Zulip model is that group PMs and 1:1 PMs are very
much alike, and the Zulip API generally treats them alike.  So
let's move toward more consistently handling them alike, and only
having conditionals to distinguish between them when we actually
want to do something different.  As a first step, in this commit,
factor out the common logic from both "constructors" and turn them
into trivial wrappers.