fetch messages: Set LAST_MESSAGE_ANCHOR to 10000000000000000 (10^16).
Fixes: #3654.

The server correctly interprets this special number as indicating
that we want the latest messages.  The larger value we had been
using will certainly get the latest messages, because it's larger
than any possible message ID, but the server won't make that
inference and can incorrectly report `found_newest: false`.

Later, we can conditionalize on the server version to use the
updated API merged in https://github.com/zulip/zulip/pull/13747,
which uses the string 'newest' instead of a special number.

[greg: revised commit message]