types [nfc]: Add convenience type MessageLike.
Replace the hack in `Outbox` (which typed it as having certain fields
which it never possessed in practice) with a slightly more
sophisticated hack in the form of a compatible auxiliary type.

This fixes a minor typing error (values of type `Outbox` could
incorrectly have been assigned the value `undefined` in those fields).
Perhaps more importantly, it also makes `Outbox` a `JSONable` type.