anchor [nfc]: Rename `constants` module to `anchor`.
We have plenty of constants that are defined in other modules
corresponding to what they're about, which is more natural than it'd
be to concentrate all constants in a single file.

The two constants that are in this file are both about the same
subject matter, namely "anchors" that define places in the message
list, so move them to a file for that purpose.  This also provides a
natural home for things we're about to add related to anchors.