msglist [nfc]: Rename a helper to getFirstUnreadIdInNarrow.
This is rather more explicit about what this selector actually means.

To the extent that "anchor" conveys any meaning of its own, it's an
answer to a UI question of how the message list should behave.  And
that's a question whose answer should live in the MessageList UI code,
not in a selectors module.