api types: Add jsdoc about CrossRealmBot; move UserOrBot into api tree.
This is a pretty confusing area of terminology in the Zulip data
model, so good to write down explicitly.

The UserOrBot union isn't itself used in the API bindings code, but
seems like a natural thing that most any client using these bindings
would need -- and I think having it around is helpful context for
the relationship between User and CrossRealmBot.