outbox [nfc]: Confirm a conjecture, to use `narrow` a bit less.
As the TODO implicitly suggests, the condition that really says what
we mean here is `item.type === private`.

And indeed that is exactly equivalent to the existing code.  This
code gets the Outbox objects to act on from the Redux store at
`state.outbox`.  That's populated entirely by MESSAGE_SEND_START
actions, which are dispatched only from `addToOutbox`.  And those
get their `type` from `extractTypeToAndSubjectFromNarrow`, and store
as `narrow` the same `Narrow` value they passed there.  That will
produce `type: 'private'` just if `isPrivateOrGroupNarrow(narrow)`,
so the latter condition is equivalent to the former.