APNs: log invalid device tokens, rather than sending them
APNs device tokens, as received through out iOS notifications library,
should be a hex string. Unfortunately, they're sometimes being sent to
the server as "[Object object]".

This is most likely due to the argument of `handleDeviceToken`
actually being an object, despite its typing as `string`, and being
improperly stringified by `encodeParamsForUrl`. However, it may
alternatively be the case that someone is mis-stringifying the object
before we get our hands on it, so log that case too.

This is work toward the resolution of #3672.