search: Make all errors clear "still fetching" state the same way.
If a query fails at this stage, before there's even a promise to
await, it's true that that's not good -- that it's a bug.

The only effect of this separation, though, is that if that happens
to your last query then we'll keep saying `isFetching: true`, even
though no work is now happening.  That's not the right behavior;
better to clear `isFetching`, just like we do if the network request
fails.  Doing so makes the code simpler and more idiomatic, too.

(The live effect of this is reduced because currently we use
`isFetching` only if we have no messages.  Still, that can happen if
the last query that succeeded had no results, or if simply no
queries have yet succeeded.)