accounts reducer: Don't clobber zulipVersion on LOGOUT or LOGIN_SUCCESS.
Spread state[index] and clobber desired properties on that, instead
of only including the desired properties.

For LOGOUT, that's apiKey and ackedPushToken.

For LOGIN_SUCCESS, that's email, apiKey, and ackedPushToken.

ACCOUNT_SWITCH and the others were not doing this.

Also make shouldHaveZulipVersion in makeAccount default to true,
adjust two callers accordingly (with the parent, a strategy to
minimize diffs for review).