i18n [nfc]: Make the semi-mechanical changes needed for new languages.
For the libdef in flow-typed/, I reran the command in the comment
at the top.

For the other two, I copied changes from a draft branch of Ray's (now
PR #3915) which includes some scripts to fully automate generating
all these files.  I dropped parts of that branch's changes where it
fixed naming so that we can, e.g., offer both zh-Hans and zh-Hant --
we should do that, but it'll need a migration for existing users who
have `zh` in their settings, so it's not as completely trivial as the
rest of these changes.

Co-authored-by: Ray Kraesig <rkraesig@zulipchat.com>