search [nfc]: Rename a method to fetchSearchMessages.
I think the word "raw" here was meant to point at the distinction,
in comparison with the handleQueryChange* methods, that this method
just does the request and doesn't manage the component's state about
the outstanding requests and their results.  We don't often say
"raw" for other things, though, and on rereading it sounds much too
strong -- like it's going to operate at a layer a few levels raw-er
than it actually turns out that it does.

Say "fetch" instead, which is a term we use in several other places
to mean something a lot like this function's job -- notably in
src/api/ and on the functions in fetchActions.js that, like this
method, go and fetch some messages from the server.