status bar [nfc]: Fix connectFlowFixMe.
The thing that had made it gnarly to use our well-typed `connect`
here was that we both took `backgroundColor` in as a prop of the
outer component and then returned a `backgroundColor` from the
`mapStateToProps` callback, to override the former as a prop on
the inner component.

The same fact is also a source of complexity and surprise when
trying to read the code as a human, because we're giving the same
name to two slightly different things right next to each other.
Doubly so because those two things are a prop of a component passed
to `connect` and the same-named prop of the resulting component --
and it's routine across the rest of our codebase that when a prop
exists on both, it's passed straight through verbatim.

So, disentangle that by introducing a distinct name.  And then
lift the fixme!

This whole "title background color" logic is still rather tanglier
than I'd like: in how `narrow` and a forced `backgroundColor` are
implicitly mutually exclusive as props to `ZulipStatusBar`; in the
use of DEFAULT_TITLE_BACKGROUND_COLOR aka `'transparent'` both as a
value with its own substantive meaning and as a sentinel meaning "no
stream here"; in how this component and `ChatNavBar` share this bit
of UI logic, but under the covers via a putative selector.  It'd be
good to disentangle those.  But now at least we have types here.