getMessages [nfc]: Restructure getMessages arguments.
`getMessages` expects a large number of arguments, which can cause
confusion, at the call site, to see which argument is for what
parameter.

Refactor getMessages to accept two arguments - the first being Auth
and the second an object that contains the parameters of the API
request, making code more readable at its call sites.