nav bar styles [nfc]: Inline the only three uses of styles.navBar.
The three uses are different enough to motiviate inlining these
styles; the next commits in this series add a loading banner to be
used in ChatNavBar but not ModalNavBar or ModalSearchNavBar.

In the child, we help along the deprecation of `context.styles`, as
in, e.g., a606bf3040, by using the ThemeContext context type.