recipient [nfc]: Check that getRecipientsIds has ownEmail when it needs it.
It's kind of worrying that this parameter is optional, because this
function relies on it in order to get a consistent answer.

Fortunately, studying the function's two call sites, we leave this
out only in a case where the message is a group PM and so we aren't
going to use it.  Make that assumption explicit, though, by checking
and throwing an error.