redux [nfc]: Fix mentions of "store" that mean the Redux "state".
The Redux store is a long-lived, mutable object.  The Redux state is
an immutable object, which the store has as a property, and which is
constantly being replaced by a new state object.  See upstream docs:
  https://redux.js.org/basics/store/

The names are confusingly similar, but we can do our bit to avoid
contributing to the confusion.