backoff machine: Add jitter.
The benefits of "jitter" are large, and the implementation is not
complex. With plain capped exponential backoff, if there is a burst
of client requests, they will all back off together, resulting in
more bursts, with gaps in between. Introducing randomness,
particularly a lot of randomness, smooths out those bursts.

Greg refers to this article:
https://aws.amazon.com/blogs/architecture/exponential-backoff-and-jitter/
and he discusses the advantages of "jitter" in more detail at
https://chat.zulip.org/#narrow/stream/92-learning/topic/exponential.20backoff.20--.20with.20jitter.
There, he claims that the work to be done with "full jitter" is O(N log N),
a great improvement over O(N^2) without any jitter.