notifications [nfc]: Further edits to `Notification` type's jsdoc.
A previous version of this jsdoc had to explain that the real type
describing the values we give this type to would be (a union of)
inexact object types, not these exact object types, because the
values can in reality have additional properties.

Happily, that's no longer true!  So there's no longer a need to talk
of "exact", vs. inexact, in explaining these types.

Also make a fewer smaller edits.  We don't interact with JSON in
sending data from our Android-native code; if JSON is involved, it's
a deeply buried RN implementation detail.  The instruction to keep
this type definition in sync with that code is important; put it back.