presence: Fix timestamp being Unix zero (50 years ago).
This is a user-visible symptom, but it's kind of a rare case that we
wouldn't expect to see normally. Sometimes, a user is displayed to
be "last active" 50 years ago, which is bogus. It turns out that it
only happens when a presence event is sent from the server (meaning
it should be as up-to-date as possible; we wouldn't expect it to be
more than a few seconds off), but it's more than 140 seconds old.

Presence events coming from the server don't have the `aggregated`
field, like presence objects do in the /register response and
presence report responses. So we have our own function,
getAggregatedPresence, to calculate it for us. This is analogous to
the status_from_timestamp function in the web app prior to 1ae07b93d
(at which point the web app made some changes to its data layer to
translate presence events to a different format, but the shape of
presence events didn't change. This was part of its handling of the
"slim presence" server capability, which affects the shape of
responses to presence pings).

As suggested in a previous commit in this series, a bug seems to
have been introduced into this function by a too-terse use of
`.reduce`, when a `.forEach` would work perfectly well.

So, to fix:

1) Use .forEach instead of reduce (that's that previous commit).

2) Assume that there is at least one ClientPresence in `presence`,
and that its timestamp is sane. This should be handled by the
server; it seems reasonable that no event would have been emitted if
there weren't sane data to transmit. But we check for this, starting
at a previous commit in this series.

3) Make the returned timestamp be whatever the latest timestamp is,
out of all the ClientPresence objects.

4) Maintain `status` as the greatest status reported within
OFFLINE_THRESHOLD_SECS (with `active` > `idle` > `offline`), or
`offline` if none was within the threshold, as explained in the
comment and implemented in the web app.

Note that (4) surfaced a behavior that looks like it might not be
fully correct, in the web app, but here, we mimic it anyway, which
prompted changing a test case and updating the comment:

The `timestamp` and the `status` fields of the returned object won't
necessarily be sourced from the same ClientPresence object from the
input. Take an example where one object has status `idle`, and
another has status `active`, and both timestamps are within
OFFLINE_THRESHOLD_SECS, but the `idle` one's timestamp is later than
the `active` one's. In that case, the aggregated `status` would be
`active`, but the timestamp would be the timestamp from the `idle`
object. This may or may not be exactly what we want; the information
conveyed is inaccurate.

Further investigation or broader refactors were avoided because of
https://github.com/zulip/zulip/issues/13734, "Improve presence
system's scalability for large realms", which signals substantial
future changes to the Presence API. See also the discussion at
https://chat.zulip.org/#narrow/stream/3-backend/topic/presence.20API/near/811593.

Fixes: #2834