eslint [nfc]: Give comments more room to breathe.
The comments on some rules and groups of rules are packed in here
pretty tight.  Promote the section-heading comments to a bolder style,
making room for the more specific comments to spread out; and blank
lines to separate more of our small groups of related rules, without
blurring the higher-level groupings.