presence [nfc]: Cut where we were passing an ignored parameter.
We stopped accepting this parameter at `doCallback` in 9e3d52dcd,
but at this indirect callsite via `setInterval` we were still
passing it.  Cut it here too.

It's a little puzzling a priori that Flow didn't complain about
this.  The cause doesn't seem to be that `setInterval` has an
inadequate type; its type in the libdef looks sensible.  But in
general, even though Flow will complain when you directly *call*
a function with more arguments than it actually takes:

  (() => {})(4); // error

it allows the function to be cast to one that takes more arguments:

  (() => {}: (number) => void);    // OK
  (() => {}: (number) => void)(4); // still OK

and similarly to be used as a value of such a type when e.g. passed
to some other function.

That is sound, so I guess it's hard to complain too much.  And it
sure is common in JS APIs to take callbacks that are passed more
arguments than most use cases will ever want to mention.  So even
though in this case I'd have liked a type error (or warning), I'm
not sure what rule I'd actually want Flow to have for doing that.