docs/contributing: Drop mention of zulipbot.
We don't in fact have zulipbot enabled on this repo; for background,
see #3831 and previous discussions linked from there.

But we've had this instruction in our CONTRIBUTING file, telling
people to invoke it and expect it to respond!  Quite reasonably, that
leaves newcomers confused when it doesn't work, as e.g. on #3975.

Fix the instruction to match our current practice.  We might enable
zulipbot in the future -- see again #3831 -- and if we do, we can
edit the instruction again accordingly.