emoji search: Leave the non-prefix matches in plain lex order.
From the user perspective, an ordering by how many characters into
the name their query appears is unhelpful for browsing, and is
likely to feel simply unpredictable and random.  Instead, keep the
emoji in familiar lexicographic order, within just the two groups of
those that match at the beginning and those that match elsewhere.

Happily this doesn't require any material amount more complexity
to implement than the other ordering does.

Also move the details of the ordering into an implementation comment.
Callers shouldn't be relying on those; but there's another important
fact they should rely on, which is that the ordering is a good one
for offering to the user to select from.  Add that promise to the
jsdoc in place of the other one.