ios build: Turn on build parallelization.
Greg notes a lack of good documentation about this feature; there is
this a WWDC talk:
https://developer.apple.com/videos/play/wwdc2018/408/?time=115.

To paraphrase Greg at #3987, it appears that the unit of
parallelization is the "target", so this could be beneficial for us
if there are a lot of targets. It turns out that there are, now that
we're using CocoaPods. In the Xcode UI, the Pods project shows many
items under "Targets": one for each line in the Podfile.

So, enable it, with a plan to disable if we ever encounter errors
that can be traced to an attempt to build two things in parallel,
when they shouldn't be.

The speed increase will be strongly influenced by the number of
cores one's CPU has.