deps: Lock rn-fetch-blob@0.10.15; breaking change at 0.10.16.
See the warning and apology for this in their README at
https://github.com/joltup/rn-fetch-blob#version-compatibility-warning.

Specifically, at
https://github.com/joltup/rn-fetch-blob/commit/0f6c3e3cc (released
in rn-fetch-blob@0.10.16), the Podspec was updated to match RN's new
pattern (added in
https://github.com/facebook/react-native/commit/2321b3fd7, released
in react-native@0.60.0) of having separate podspecs for each Xcode
project, instead of using subspecs for them.

We can allow upgrading to 0.10.16 when we switch to
react-native@0.60.0 (that's open as #3548).